+++ trunk/sfox/src/starengine/engine/bbox.h
@@ -9,7 +9,7 @@ typedef struct bbox2d {
 } bbox2d;
 
 /*tl=top left br=bottom right*/
-extern void bbox2d_set(bbox2d *bb, double tlx, double tly, double brx, double bry);
+extern void bbox2d_set(bbox2d *bb, float tlx, float tly, float brx, float bry);
 extern unsigned int bbox2d_width(bbox2d *bb);
 extern unsigned int bbox2d_height(bbox2d *bb);
